ECE651-104: Resolving sonarCloud issues
