+++ src/components/SingleUserOrderDetail.js
@@ -1,12 +1,35 @@
 import React from "react";
+import firebase from '../config/firebase';
+import { useSelector } from "react-redux";
+
+function handleSendToInProgressBtn(userUid, orderId, restaurantUid, status) {
+  console.log("user id:", userUid);
+  console.log("order id:", orderId);
+  console.log("restaurant id:", restaurantUid);
+  firebase.firestore().collection('users').doc(restaurantUid).collection('orderRequest').doc(orderId).update({
+      status: status,
+  }).then(() => {
+      console.log("First Seccussfully send to IN PROGRESS")
+      firebase.firestore().collection('users').doc(userUid).collection('myOrder').doc(orderId).update({
+          status: status,
+      }).then(()=>{
+          console.log("Second Seccussfully send to IN PROGRESS")
+      })
+  })
+};
 
 export default function SingleUserOrderDetail({
+  orderId,
+  userUid,
   order_status,
   order_status_color,
   restaurant_name,
   total_price,
   orderItemList,
+  nextaction,
+  status
 }) {
+  const restaurantUid = useSelector((state) => state.loggedInUser.userId);
   return (
     <div>
       <div className="flex">
@@ -33,7 +56,14 @@ export default function SingleUserOrderDetail({
         );
       })}
 
-      <div className="w-full my-2 flex">
+      <div className="mt-16 w-full my-2 flex">
+        {
+          (status === "") ? 
+          (<div></div>) : 
+          (
+            <button type="button" className="ml-2 rounded-lg bg-orange p-2.5 text-sm font-medium text-black" onClick={() => handleSendToInProgressBtn(userUid, orderId, restaurantUid, status)}>{nextaction}</button>
+          )
+        }
         <div className="ml-auto">Total :${total_price}</div>
       </div>
